Reorganized source/header files for http code architecture
