--- include/http/http_parse.h
@@ -1,8 +1,0 @@
-//
-// Created by jordan on 2021-11-08.
-//
-
-#ifndef IBEACON_HTTP_PARSE_H
-#define IBEACON_HTTP_PARSE_H
-
-#endif //IBEACON_HTTP_PARSE_H
