--- src/http/http_parse.c
@@ -1,6 +1,0 @@
-//
-// Created by jordan on 2021-11-08.
-//
-
-#include "http/http_parse.h"
-#define Hello 9
